Add ability to edit your failed BST post
In the event that a BST post failed validation the user would not be able to repost it for 6 hours (our current slowmode setting).  To get around this we now warn the user and set a 30 minute timer for them to update their post.

Once the post is updated it gets revalidated